More revisions to classifier, moving of count_entities function to different file for convenience.
